backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brikell-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "openai>=1.40",
    "tenacity>=8.0",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "pandas>=2.2",
    "tabulate>=0.9",
    "weasyprint>=62",
    "jinja2>=3.1",
]

[tool.setuptools.packages.find]
include = ["app*"]
